Se renombran clases
