+++ src/main/java/paqueteTurismoTM/App.java
@@ -1,0 +1,12 @@
+package paqueteTurismoTM;
+
+import java.io.IOException;
+
+public class App {
+
+	public static void main(String[] args) throws IOException {
+		TurismoTM boleteria = new TurismoTM();
+		boleteria.sugerenciaCliente();
+	}
+
+}
